backend-raw/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-ninja==1.4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
inflection==0.5.1
loguru==0.7.3
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.2.1
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
